Removing hardcode
