+++ HW6_Room/app/src/main/java/com/example/hw6_room/screens/bottomNavigation/FavoriteMemesFragment.kt
@@ -1,4 +1,4 @@
-package com.example.hw6_room.screens
+package com.example.hw6_room.screens.bottomNavigation
 
 import android.os.Bundle
 import android.util.Log
@@ -6,7 +6,6 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
@@ -14,19 +13,17 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestManager
 import com.example.hw6_room.MainActivity
 import com.example.hw6_room.R
-import com.example.hw6_room.databinding.FragmentMainScreenBinding
+import com.example.hw6_room.databinding.FragmentFavoriteMemesBinding
 import com.example.hw6_room.db.entity.MemeEntity
-import com.example.hw6_room.db.repository.MemeRepository
 import com.example.hw6_room.di.ServiceLocator
 import com.example.hw6_room.recyclerView.MemesAdapter
 import com.example.hw6_room.utils.Constants
 import com.example.hw6_room.utils.SessionManager
 import kotlinx.coroutines.launch
 
+class FavoriteMemesFragment : Fragment() {
 
-class MainScreenFragment : Fragment() {
-
-    private var viewBinding: FragmentMainScreenBinding? = null
+    private var viewBinding: FragmentFavoriteMemesBinding? = null
     private var memesAdapter: MemesAdapter? = null
     private val memeRepository = ServiceLocator.getMemeRepository()
     private lateinit var sessionManager: SessionManager
@@ -35,7 +32,7 @@ class MainScreenFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        viewBinding = FragmentMainScreenBinding.inflate(inflater, container, false)
+        viewBinding = FragmentFavoriteMemesBinding.inflate(inflater, container, false)
         return viewBinding?.root
     }
 
@@ -45,53 +42,37 @@ class MainScreenFragment : Fragment() {
         (activity as MainActivity).setBottomNavigationVisibility(true)
         val glide = Glide.with(requireContext())
         initRecyclerView(glide)
-        initFab()
     }
 
     private fun initRecyclerView(requestManager: RequestManager) {
         lifecycleScope.launch {
             val userId = sessionManager.getUserId()
-            val memesList = memeRepository.getUserMemes(userId).toMutableList()
+            val favoriteMemesList = memeRepository.getFavoriteMemes(userId).toMutableList()
             memesAdapter = MemesAdapter(
-                list = memesList,
+                list = favoriteMemesList,
                 requestManager = requestManager,
                 onItemClick = { memeId ->
                     openMemeDetails(memeId)
                 },
                 onItemLongClick = { meme, position ->
-                    showDeleteConfirmationDialog(meme, position)
+                    toggleFavorite(meme, position)
                 },
                 onFavoriteClick = { position, _ ->
-                    toggleFavorite(memesList[position - 1], position)
+                    if (position in favoriteMemesList.indices) {
+                        toggleFavorite(favoriteMemesList[position], position)
+                    } else {
+                        val errorMessage = getString(R.string.invalid_meme_index, position, favoriteMemesList.size)
+                        Log.e(Constants.FAVORITE_MEMES_FRAGMENT_TAG, errorMessage)
+                    }
                 }
+
             )
             viewBinding?.run {
-                mainRecycler.adapter = memesAdapter
-                mainRecycler.layoutManager = GridLayoutManager(requireContext(), 2)
+                favoriteRecycler.adapter = memesAdapter
+                favoriteRecycler.layoutManager = GridLayoutManager(requireContext(), 2)
             }
         }
     }
-
-    private fun deleteMeme(meme: MemeEntity, position: Int) {
-        lifecycleScope.launch {
-            memeRepository.deleteMeme(meme.id, sessionManager.getUserId())
-            memesAdapter?.removeItem(position)
-        }
-    }
-
-    private fun showDeleteConfirmationDialog(meme: MemeEntity, position: Int) {
-        val dialog = AlertDialog.Builder(requireContext())
-            .setTitle(getString(R.string.delete_confirmation_title))
-            .setMessage(getString(R.string.delete_confirmation_message))
-            .setPositiveButton(getString(R.string.delete_button)) { _, _ ->
-                deleteMeme(meme, position)
-            }
-            .setNegativeButton(getString(R.string.cancel_button), null)
-            .create()
-
-        dialog.show()
-    }
-
 
     private fun toggleFavorite(meme: MemeEntity, position: Int) {
         lifecycleScope.launch {
@@ -105,16 +86,7 @@ class MainScreenFragment : Fragment() {
         val bundle = Bundle().apply {
             putInt(Constants.MEME_DETAILS_BUNDLE_KEY, memeId)
         }
-        findNavController().navigate(R.id.action_memesListFragment_to_memeDetailsFragment, bundle)
-    }
-
-
-    private fun initFab() {
-        viewBinding?.run {
-            fab.setOnClickListener {
-                findNavController().navigate(R.id.action_mainScreenFragment_to_addContentFragment)
-            }
-        }
+        findNavController().navigate(R.id.action_favoriteMemesFragment_to_memeDetailsFragment, bundle)
     }
 
 
@@ -122,7 +94,4 @@ class MainScreenFragment : Fragment() {
         super.onDestroyView()
         viewBinding = null
     }
-
-
-
 }
