+++ HW6_Room/app/src/main/java/com/example/hw6_room/screens/bottomNavigation/MemesListFragment.kt
@@ -1,11 +1,11 @@
-package com.example.hw6_room.screens
+package com.example.hw6_room.screens.bottomNavigation
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
@@ -13,7 +13,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestManager
 import com.example.hw6_room.MainActivity
 import com.example.hw6_room.R
-import com.example.hw6_room.databinding.FragmentFavoriteMemesBinding
+import com.example.hw6_room.databinding.FragmentMainScreenBinding
 import com.example.hw6_room.db.entity.MemeEntity
 import com.example.hw6_room.di.ServiceLocator
 import com.example.hw6_room.recyclerView.MemesAdapter
@@ -21,9 +21,10 @@ import com.example.hw6_room.utils.Constants
 import com.example.hw6_room.utils.SessionManager
 import kotlinx.coroutines.launch
 
-class FavoriteMemesFragment : Fragment() {
 
-    private var viewBinding: FragmentFavoriteMemesBinding? = null
+class MemesListFragment : Fragment() {
+
+    private var viewBinding: FragmentMainScreenBinding? = null
     private var memesAdapter: MemesAdapter? = null
     private val memeRepository = ServiceLocator.getMemeRepository()
     private lateinit var sessionManager: SessionManager
@@ -32,7 +33,7 @@ class FavoriteMemesFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        viewBinding = FragmentFavoriteMemesBinding.inflate(inflater, container, false)
+        viewBinding = FragmentMainScreenBinding.inflate(inflater, container, false)
         return viewBinding?.root
     }
 
@@ -42,37 +43,53 @@ class FavoriteMemesFragment : Fragment() {
         (activity as MainActivity).setBottomNavigationVisibility(true)
         val glide = Glide.with(requireContext())
         initRecyclerView(glide)
+        initFab()
     }
 
     private fun initRecyclerView(requestManager: RequestManager) {
         lifecycleScope.launch {
             val userId = sessionManager.getUserId()
-            val favoriteMemesList = memeRepository.getFavoriteMemes(userId).toMutableList()
+            val memesList = memeRepository.getUserMemes(userId).toMutableList()
             memesAdapter = MemesAdapter(
-                list = favoriteMemesList,
+                list = memesList,
                 requestManager = requestManager,
                 onItemClick = { memeId ->
                     openMemeDetails(memeId)
                 },
                 onItemLongClick = { meme, position ->
-                    toggleFavorite(meme, position)
+                    showDeleteConfirmationDialog(meme, position)
                 },
                 onFavoriteClick = { position, _ ->
-                    if (position in favoriteMemesList.indices) {
-                        toggleFavorite(favoriteMemesList[position], position)
-                    } else {
-                        val errorMessage = getString(R.string.invalid_meme_index, position, favoriteMemesList.size)
-                        Log.e(Constants.FAVORITE_MEMES_FRAGMENT_TAG, errorMessage)
-                    }
+                    toggleFavorite(memesList[position - 1], position)
                 }
-
             )
             viewBinding?.run {
-                favoriteRecycler.adapter = memesAdapter
-                favoriteRecycler.layoutManager = GridLayoutManager(requireContext(), 2)
+                mainRecycler.adapter = memesAdapter
+                mainRecycler.layoutManager = GridLayoutManager(requireContext(), 2)
             }
         }
     }
+
+    private fun deleteMeme(meme: MemeEntity, position: Int) {
+        lifecycleScope.launch {
+            memeRepository.deleteMeme(meme.id, sessionManager.getUserId())
+            memesAdapter?.removeItem(position)
+        }
+    }
+
+    private fun showDeleteConfirmationDialog(meme: MemeEntity, position: Int) {
+        val dialog = AlertDialog.Builder(requireContext())
+            .setTitle(getString(R.string.delete_confirmation_title))
+            .setMessage(getString(R.string.delete_confirmation_message))
+            .setPositiveButton(getString(R.string.delete_button)) { _, _ ->
+                deleteMeme(meme, position)
+            }
+            .setNegativeButton(getString(R.string.cancel_button), null)
+            .create()
+
+        dialog.show()
+    }
+
 
     private fun toggleFavorite(meme: MemeEntity, position: Int) {
         lifecycleScope.launch {
@@ -86,7 +103,16 @@ class FavoriteMemesFragment : Fragment() {
         val bundle = Bundle().apply {
             putInt(Constants.MEME_DETAILS_BUNDLE_KEY, memeId)
         }
-        findNavController().navigate(R.id.action_favoriteMemesFragment_to_memeDetailsFragment, bundle)
+        findNavController().navigate(R.id.action_memesListFragment_to_memeDetailsFragment, bundle)
+    }
+
+
+    private fun initFab() {
+        viewBinding?.run {
+            fab.setOnClickListener {
+                findNavController().navigate(R.id.action_mainScreenFragment_to_addContentFragment)
+            }
+        }
     }
 
 
@@ -94,4 +120,7 @@ class FavoriteMemesFragment : Fragment() {
         super.onDestroyView()
         viewBinding = null
     }
+
+
+
 }
