+++ HW6_Room/app/src/main/java/com/example/hw6_room/screens/entry/LoginFragment.kt
@@ -76,7 +76,6 @@ class LoginFragment : Fragment() {
 
     override fun onDestroyView() {
         super.onDestroyView()
-        (activity as MainActivity).setBottomNavigationVisibility(true)
         viewBinding = null
     }
 }
