+++ HW6_Room/app/src/main/res/layout/fragment_main_screen.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.MainScreenFragment">
+    tools:context=".screens.bottomNavigation.MemesListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/main_recycler"
