+++ HW6_Room/app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".screens.SettingsFragment">
+    tools:context=".screens.bottomNavigation.SettingsFragment">
 
     <TextView
         android:id="@+id/textView3"
